bcachefs: fix truncate without a size change
Do not attempt to shortcut a truncate when the given new size is
the same as the current size. There may be blocks allocated to the
file that extend beyond the i_size. The ctime and mtime should
not be updated in this case.

Signed-off-by: Dan Robertson <dan@dlrobertson.com>
(cherry picked from commit b8336ebba9d924056ce2005de90737f13b9e3cb1)

 Conflicts:
	fs/bcachefs/fs-io.c